refactor(acp): extract terminal subsystem into terminalManager.ts
Moves the 9 ACP terminal handlers + the 4 per-session terminal-state maps off
AcpStructuredSession into a new AcpTerminalManager, injected via a context
{ projectLocation, cwd, assertRequestSession }. session.ts drops 1583->1425.
The records/env helpers (sessionTerminal.ts, sessionTerminalLaunch.ts) were
already extracted; only the handler bodies moved. Manager never imports session
(one-directional). Behavior-preserving; session.test.ts (62) +
canonicalMapping.test.ts (47) green.
